requirements.txt
django
djangorestframework
djangorestframework-simplejwt
channels
psycopg2-binary
pillow
python-decouple